sql: remove legacy FK code
This change removes all the legacy code around FK checks and cascades, and
anything that becomes unused as a result.

Release note: None
