+++ pkg/sql/row/deleter.go
@@ -26,8 +26,6 @@ type Deleter struct {
 	Helper               rowHelper
 	FetchCols            []sqlbase.ColumnDescriptor
 	FetchColIDtoRowIndex map[sqlbase.ColumnID]int
-	Fks                  fkExistenceCheckForDelete
-	cascader             *cascader
 	// For allocation avoidance.
 	key roachpb.Key
 }
@@ -42,60 +40,7 @@ func MakeDeleter(
 	txn *kv.Txn,
 	codec keys.SQLCodec,
 	tableDesc *sqlbase.ImmutableTableDescriptor,
-	fkTables FkTableMetadata,
 	requestedCols []sqlbase.ColumnDescriptor,
-	checkFKs checkFKConstraints,
-	evalCtx *tree.EvalContext,
-	alloc *sqlbase.DatumAlloc,
-) (Deleter, error) {
-	rowDeleter, err := makeRowDeleterWithoutCascader(
-		ctx, txn, codec, tableDesc, fkTables, requestedCols, checkFKs, alloc,
-	)
-	if err != nil {
-		return Deleter{}, err
-	}
-	if checkFKs == CheckFKs {
-		var err error
-		rowDeleter.cascader, err = makeDeleteCascader(ctx, txn, tableDesc, fkTables, evalCtx, alloc)
-		if err != nil {
-			return Deleter{}, err
-		}
-		// If we are performing a cascade operation for a particular foreign
-		// key constraint, we don't also need to perform a foreign key
-		// existence check after the delete for the same foreign key
-		// constraint. This pass removes unnecessary existence helpers.
-		// In particular, we omit checks for CASCADE and SET NULL because
-		// after the cascader has finished deleting or setting rows to
-		// NULL, we don't need to verify the result of those operations.
-		// TODO (rohany): This code will be removed once the optimizer
-		//  handles cascade operations.
-		for k, helpers := range rowDeleter.Fks.fks {
-			index := 0
-			for i := range helpers {
-				helper := &helpers[i]
-				if helper.ref.OnDelete == sqlbase.ForeignKeyReference_CASCADE ||
-					helper.ref.OnDelete == sqlbase.ForeignKeyReference_SET_NULL {
-					continue
-				}
-				helpers[index] = *helper
-				index++
-			}
-			rowDeleter.Fks.fks[k] = helpers[:index]
-		}
-	}
-	return rowDeleter, nil
-}
-
-// makeRowDeleterWithoutCascader creates a rowDeleter but does not create an
-// additional cascader.
-func makeRowDeleterWithoutCascader(
-	ctx context.Context,
-	txn *kv.Txn,
-	codec keys.SQLCodec,
-	tableDesc *sqlbase.ImmutableTableDescriptor,
-	fkTables FkTableMetadata,
-	requestedCols []sqlbase.ColumnDescriptor,
-	checkFKs checkFKConstraints,
 	alloc *sqlbase.DatumAlloc,
 ) (Deleter, error) {
 	indexes := tableDesc.DeletableIndexes()
@@ -138,13 +83,6 @@ func makeRowDeleterWithoutCascader(
 		FetchCols:            fetchCols,
 		FetchColIDtoRowIndex: fetchColIDtoRowIndex,
 	}
-	if checkFKs == CheckFKs {
-		var err error
-		if rd.Fks, err = makeFkExistenceCheckHelperForDelete(ctx, txn, codec, tableDesc, fkTables,
-			fetchColIDtoRowIndex, alloc); err != nil {
-			return Deleter{}, err
-		}
-	}
 
 	return rd, nil
 }
@@ -154,7 +92,7 @@ func makeRowDeleterWithoutCascader(
 // orphaned rows. The bytesMonitor is only used if cascading/fk checking and can
 // be nil if not.
 func (rd *Deleter) DeleteRow(
-	ctx context.Context, b *kv.Batch, values []tree.Datum, checkFKs checkFKConstraints, traceKV bool,
+	ctx context.Context, b *kv.Batch, values []tree.Datum, traceKV bool,
 ) error {
 
 	// Delete the row from any secondary indices.
@@ -201,24 +139,6 @@ func (rd *Deleter) DeleteRow(
 		rd.key = nil
 	}
 
-	if rd.cascader != nil {
-		if err := rd.cascader.cascadeAll(
-			ctx,
-			rd.Helper.TableDesc,
-			tree.Datums(values),
-			nil, /* updatedValues */
-			rd.FetchColIDtoRowIndex,
-			traceKV,
-		); err != nil {
-			return err
-		}
-	}
-	if rd.Fks.checker != nil && checkFKs == CheckFKs {
-		if err := rd.Fks.addAllIdxChecks(ctx, values, traceKV); err != nil {
-			return err
-		}
-		return rd.Fks.checker.runCheck(ctx, values, nil)
-	}
 	return nil
 }
 
@@ -227,14 +147,6 @@ func (rd *Deleter) DeleteRow(
 func (rd *Deleter) DeleteIndexRow(
 	ctx context.Context, b *kv.Batch, idx *sqlbase.IndexDescriptor, values []tree.Datum, traceKV bool,
 ) error {
-	if rd.Fks.checker != nil {
-		if err := rd.Fks.addAllIdxChecks(ctx, values, traceKV); err != nil {
-			return err
-		}
-		if err := rd.Fks.checker.runCheck(ctx, values, nil); err != nil {
-			return err
-		}
-	}
 	// We want to include empty k/v pairs because we want
 	// to delete all k/v's for this row. By setting includeEmpty
 	// to true, we will get a k/v pair for each family in the row,
