+++ pkg/sql/sqlbase/structured.go
@@ -3039,28 +3039,6 @@ func (desc *TableDescriptor) FindFKByName(name string) (*ForeignKeyConstraint, e
 	return nil, fmt.Errorf("fk %q does not exist", name)
 }
 
-// FindFKForBackRef searches the table descriptor for the foreign key constraint
-// that matches the supplied backref, which is present on the supplied table id.
-// Our current restriction that each column in a table can be on the referencing
-// side of at most one FK relationship means that there's no possibility of
-// ambiguity when finding the forward FK reference for a given backreference,
-// since ambiguity would require an identical list of referencing columns. This
-// would continue to hold even if we were to lift that restriction (see #38850), as long as
-// it were still prohibited to create multiple foreign key relationships with
-// exactly identical lists of referenced and referencing columns, which we think
-// is a reasonable restriction (even though Postgres does allow doing this).
-func (desc *TableDescriptor) FindFKForBackRef(
-	referencedTableID ID, backref *ForeignKeyConstraint,
-) (*ForeignKeyConstraint, error) {
-	for i := range desc.OutboundFKs {
-		fk := &desc.OutboundFKs[i]
-		if fk.ReferencedTableID == referencedTableID && fk.Name == backref.Name {
-			return fk, nil
-		}
-	}
-	return nil, errors.AssertionFailedf("could not find fk for backref %v", backref)
-}
-
 // IsInterleaved returns true if any part of this this table is interleaved with
 // another table's data.
 func (desc *TableDescriptor) IsInterleaved() bool {
