+++ pkg/sql/tablewriter_delete.go
@@ -67,7 +67,7 @@ func (td *tableDeleter) row(
 	ctx context.Context, values tree.Datums, ignoreIndexes util.FastIntSet, traceKV bool,
 ) error {
 	td.batchSize++
-	return td.rd.DeleteRow(ctx, td.b, values, row.CheckFKs, traceKV)
+	return td.rd.DeleteRow(ctx, td.b, values, traceKV)
 }
 
 // deleteAllRows runs the kv operations necessary to delete all sql rows in the
